Added new exceptions module and refactored existing code to use it, removed obsolete file
